feat: Implement CommandHandler
